Added suportin Ef core and done program work by Dependency Injection

+++ CarDealership.Domain/Car.cs
@@ -6,11 +6,11 @@ namespace CarDealershipApp.Domain
 {
     public class Car
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Number { get; set; }
         public bool IsSold { get; set; }
         public Client Client { get; set; }
-        public int? ClientId { get; set; }
+        public long? ClientId { get; set; }
         public decimal Price { get; set; }
         public static Car CreateCar(string number, decimal price)
         {
